# contract/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cleared API",
    "version": "1.0.0",
    "description": "Steps 1 to 3 of a deal: signing in, deals, brief to checklist, the invite and the brand's link, change requests, agreeing and the holds. A creator's session and a brand's are HttpOnly cookies set by this API. Every error has one shape: a code and, where useful, the field it is about."
  },
...
  "paths": {
    "/auth/demo": {
      "post": {
        "summary": "Try the demo account: a form the page posts, answered with a redirect into the app (DS-FR-06)",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "next",
            "in": "query"
          }
        ],
        "responses": {
          "303": {
            "description": "Signed in to a new demo account and sent to `next` or /deals; or to /sign-in?problem=demo_limit"
          }
        }
      }
    },
...
  },
...
}